requirements.txt
channels==2
daphne==2
Django==3
django-allauth
django-mptt
Pillow

Twisted @ file:///C:/Users/alexw/projects/home/Django/Twisted-20.3.0-cp38-cp38-win32.whl
